+++ axum-extra/src/middleware/mod.rs
@@ -1,0 +1,4 @@
+//! Additional middleware.
+
+#[cfg(feature = "opentelemetry")]
+pub mod opentelemetry;